pyproject.toml

[project]
name = "ziracle"
description = "Implementation for the Ziracle method and experiments based on the Focus-CoIR dataset."
authors = [
    {name = "Erdösi, Florian", email = "Florian.Erdoesi@vector.com"},
]
dependencies = [
    "jsonlines>=3.1.0",
    "numpy>=1.21",
    "tqdm>=4.65.0",
    "Pillow>=9.5.0",
    "torch>=2.1.0",
    "salesforce-lavis==1.0.2",
    "mlflow>=2.11.3",
]
requires-python = ">=3.10,<=3.12"
readme = "README.md"
license = {file = "LICENSE"}
dynamic = ["version"]

[build-system]
requires = ["pdm-pep517>=1.0"]
build-backend = "pdm.pep517.api"

[tool.pdm]
version = { source = "scm" }

[tool.pdm.dev-dependencies]
dev = [
    "pytest>=7.3.1",
    "ipykernel>=6.23.0",
]

[tool.pdm.build]
package-dir = "src/"
includes = ["src/ziracle/",]
